[wallet] move constants to walletconstants.h
